% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{calibrate_sizes}
\alias{calibrate_sizes}
\title{Function to scale values between two given limits}
\usage{
calibrate_sizes(x, high, low)
}
\arguments{
\item{x}{Numeric vector}

\item{high}{Numeric value of largest size}

\item{low}{Numeric value of smallest size}
}
\value{
Numeric vector
}
\description{
Could be useful to modify a vector of sizes to between, say 1 and 3, before passing to `onto_plot`.
}
\examples{
calibrate_sizes(c("HP:0000001"=10, "HP:0000006"=5), high=3, low=1)
}

