% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{colour_by_frequency}
\alias{colour_by_frequency}
\title{Function to assign colours to terms based on frequency with which terms appear in \code{term_sets}}
\usage{
colour_by_frequency(ontology, terms, term_sets,
  colour_func = colorRampPalette(c("Yellow", "Green", "#0099FF")))
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{terms}{Character vector of ontological terms}

\item{term_sets}{List of character vectors of ontological term IDs}

\item{colour_func}{Function capable of returning a set of colours, given the number of colours it needs to return}
}
\value{
Character vector of colours, named by term
}
\description{
Function to assign colours to terms based on frequency with which terms appear in \code{term_sets}
}
\seealso{
\code{\link{colour_by_term_set}}, \code{\link{colour_by_population_frequency}}
}

