% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{colour_by_term_set}
\alias{colour_by_term_set}
\title{Function to set colours of nodes in plot to distinguish terms belonging to different term sets}
\usage{
colour_by_term_set(ontology, terms, term_sets, colour_generator = rainbow,
  alpha = 0.5)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{terms}{Character vector of ontological terms}

\item{term_sets}{List of character vectors of ontological term IDs}

\item{colour_generator}{Function which returns a vector of colours, e.g. \code{rainbow} or \code{heat.colors}.}

\item{alpha}{\code{alpha} parameter to pass to \code{colour_generator}.}
}
\value{
Character vector of colours, named by term.
}
\description{
Function to set colours of nodes in plot to distinguish terms belonging to different term sets
}
\seealso{
\code{\link{colour_by_frequency}}, \code{\link{colour_by_population_frequency}}
}

