% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{dot_string}
\alias{dot_string}
\title{\code{ontology_plot} object to dot string}
\usage{
dot_string(ontology_plot)
}
\arguments{
\item{ontology_plot}{Object of class `ontology_plot` to export.}
}
\value{
String
}
\description{
\code{ontology_plot} object to dot string
}
\seealso{
\code{\link{onto_plot}}
}

