% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{get_node_friendly_long_names}
\alias{get_node_friendly_long_names}
\title{Split up node labels across lines so they fit in nodes better}
\usage{
get_node_friendly_long_names(ontology, terms, official_names = FALSE)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{terms}{Character vector of ontological terms}

\item{official_names}{Logical value indicating whether to use the exact names from the ontology. Otherwise, shortened, capitalised names are used.}
}
\value{
Character vector.
}
\description{
Split up node labels across lines so they fit in nodes better
}
\examples{
library(ontologyIndex)
data(hpo)
get_node_friendly_long_names(hpo, c("HP:0001873", "HP:0011877"))
}
