% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_shortened_names}
\alias{get_shortened_names}
\title{Get human readable, shortened (where possible) ontological term names}
\usage{
get_shortened_names(ontology, terms)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{terms}{Character vector of ontological terms}
}
\value{
Character vector
}
\description{
Get human readable, shortened (where possible) ontological term names
}
\examples{
library(ontologyIndex)
data(hpo)
get_shortened_names(hpo, c("HP:0001873", "HP:0011877"))
}
