% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_grid.R
\name{grid_terms}
\alias{grid_terms}
\title{Get set of HPO terms appropriate for showing in a grid}
\usage{
grid_terms(ontology, term_sets)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{term_sets}{List of character vectors of ontological term IDs}
}
\value{
Character vector of terme IDs.
}
\description{
Get set of HPO terms appropriate for showing in a grid
}
