% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{label_by_term_set}
\alias{label_by_term_set}
\title{Function to label nodes by \code{term_set}}
\usage{
label_by_term_set(ontology, terms, term_sets)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{terms}{Character vector of ontological terms}

\item{term_sets}{List of character vectors of ontological term IDs}
}
\value{
Character vector of colours, named by term.
}
\description{
Function to label nodes by \code{term_set}
}
\seealso{
\code{\link{simple_labels}}, \code{\link{label_by_frequency}}, \code{\link{long_labels}}
}

