% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{long_labels}
\alias{long_labels}
\title{Function to assign detailed node labels to terms}
\usage{
long_labels(ontology, terms, term_sets, frequencies)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{terms}{Character vector of ontological terms}

\item{term_sets}{List of character vectors of ontological term IDs}

\item{frequencies}{Numeric vector of term frequencies named by term IDs}
}
\value{
Character vector of labels, named by term.
}
\description{
Label includes term ID, term name, number of instances of term amongst \code{term_sets} and percentage frequency in population.
}
\seealso{
\code{\link{simple_labels}}, \code{\link{label_by_frequency}}, \code{\link{label_by_term_set}}
}

