% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{official_labels}
\alias{official_labels}
\title{Get official names for terms}
\usage{
official_labels(ontology, terms)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{terms}{Character vector of ontological terms}
}
\value{
Character vector of labels, named by term.
}
\description{
Get official names for terms
}
\seealso{
\code{\link{simple_labels}}
}

