% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{only_phenotype_abnormalities}
\alias{only_phenotype_abnormalities}
\title{Remove terms not descending from phenotypic abnormality}
\usage{
only_phenotype_abnormalities(ontology, terms)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{terms}{Character vector of ontological terms}
}
\value{
Character vector.
}
\description{
Remove terms not descending from phenotypic abnormality
}
\seealso{
\code{\link{remove_terms_with_less_than_n_occurrences}}, \code{\link{n_most_frequent_terms}}
}

