% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{p_values_for_occurrence_of_term_in_group}
\alias{p_values_for_occurrence_of_term_in_group}
\title{Get p-values for observing at least as many of each term as occur in \code{term_sets} given the population frequencies of the terms}
\usage{
p_values_for_occurrence_of_term_in_group(ontology, term_sets, terms_freq)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{term_sets}{List of character vectors of ontological term IDs}

\item{terms_freq}{Numeric vector of population frequencies of terms.}
}
\value{
Numeric vector of log p-values named by correspnding term.
}
\description{
Get p-values for observing at least as many of each term as occur in \code{term_sets} given the population frequencies of the terms
}
\seealso{
\code{\link{width_by_significance}}
}

