% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot.ontology_plot}
\alias{plot.ontology_plot}
\title{Plotting function for \code{ontology_plot} object}
\usage{
\method{plot}{ontology_plot}(x, ...)
}
\arguments{
\item{x}{Object of class ontologicalPlot.}

\item{...}{Other options passed to plot().}
}
\value{
Nothing, side-effect: plots a graph.
}
\description{
Plotting function for \code{ontology_plot} object
}

