% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_grid.R
\name{plot_annotation_grid}
\alias{plot_annotation_grid}
\title{Plot a logical matrix of term annotation}
\usage{
plot_annotation_grid(..., on_colour = "#FF0000FF", off_colour = "#FFFFBFFF")
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{annotation_grid}}.}

\item{on_colour}{Colour to use to show presence of term.}

\item{off_colour}{Colour to use to show absence of term.}
}
\value{
Plots heatmap.
}
\description{
Plot a logical matrix of term annotation
}
