% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{print.ontology_plot}
\alias{print.ontology_plot}
\title{Print function for \code{ontology_plot} object}
\usage{
\method{print}{ontology_plot}(x, ...)
}
\arguments{
\item{x}{Object of class ontologicalPlot.}

\item{...}{Other options passed to be passed to plot().}
}
\value{
Nothing. Side-effect: plots graphs.
}
\description{
Print function for \code{ontology_plot} object
}

