% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{remove_links}
\alias{remove_links}
\title{Remove terms which just link two other terms}
\usage{
remove_links(ontology, terms, hard = FALSE)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{terms}{Character vector of ontological terms}

\item{hard}{Logical value indicating whether to remove alternative direct paths to leaf nodes.}
}
\value{
Character vector.
}
\description{
Remove terms which just link two other terms
}
\examples{
library(ontologyIndex)
data(hpo)
remove_links(hpo, c("HP:0001873"))
}
\seealso{
\code{\link{remove_uninformative_terms}}
}

