% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{simple_cap}
\alias{simple_cap}
\title{Capitalise words in character vector}
\usage{
simple_cap(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector
}
\description{
Capitalise words in character vector
}
\examples{
simple_cap(c("a simple test", "Another-test"))
}
