% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{simple_labels}
\alias{simple_labels}
\title{Get simplified labels for terms}
\usage{
simple_labels(ontology, terms)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{terms}{Character vector of ontological terms}
}
\value{
Character vector of labels, named by term.
}
\description{
Get simplified labels for terms
}
\seealso{
\code{\link{official_labels}}
}
