% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{to_svg_string}
\alias{to_svg_string}
\title{Convert \code{ontology_plot} to SVG string}
\usage{
to_svg_string(op)
}
\arguments{
\item{op}{Object of class \code{ontology_plot}.}
}
\value{
Character vector of length 1 containing SVG representation of node.
}
\description{
Note that by setting \code{"id"} and \code{"class"} attributes it enables nodes to be selected for manipulation using Javascript if interactivity is desired.
}
\seealso{
\code{\link{onto_plot}}, \code{\link{get_ontology_plot}}
}
