% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{width_by_frequency}
\alias{width_by_frequency}
\title{Function to get node sizes for terms based on frequency in \code{term_sets}}
\usage{
width_by_frequency(ontology, terms, term_sets)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{terms}{Character vector of ontological terms}

\item{term_sets}{List of character vectors of ontological term IDs}
}
\value{
Character vector of sizes, named by term
}
\description{
Function to get node sizes for terms based on frequency in \code{term_sets}
}
\seealso{
\code{\link{width_by_significance}}
}

