% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{width_by_significance}
\alias{width_by_significance}
\title{Function to get node sizes for terms based on statistical significance of seeing at least this number of each term in \code{term_sets}}
\usage{
width_by_significance(ontology, terms, term_sets, frequencies)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{terms}{Character vector of ontological terms}

\item{term_sets}{List of character vectors of ontological term IDs}

\item{frequencies}{Numeric vector of term frequencies named by term IDs}
}
\value{
Character vector of sizes, named by term
}
\description{
Function to get node sizes for terms based on statistical significance of seeing at least this number of each term in \code{term_sets}
}
\seealso{
\code{\link{width_by_frequency}}
}
