% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{write_dot}
\alias{write_dot}
\title{Function to write dot file}
\usage{
write_dot(ontology_plot, file)
}
\arguments{
\item{ontology_plot}{Object of class `ontology_plot` to export.}

\item{file}{Character value of target file path.}
}
\value{
Nothing, side effect - writes to file.
}
\description{
Function to write dot file
}
\seealso{
\code{\link{onto_plot}}
}

