% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{colour_by_term_set}
\alias{colour_by_term_set}
\title{Function to set colours of nodes in plot to distinguish terms belonging to different term sets}
\usage{
colour_by_term_set(ontology, terms, term_sets)
}
\arguments{
\item{ontology}{\code{ontology_index} object}

\item{terms}{Character vector of ontological terms}

\item{term_sets}{List of character vectors of ontological term IDs}
}
\value{
Character vector of colours, named by term.
}
\description{
Function to set colours of nodes in plot to distinguish terms belonging to different term sets
}
\seealso{
\code{\link{colour_by_frequency}}, \code{\link{colour_by_population_frequency}}
}

